# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Matador API
  description: Matador on-demand car rental API. Customer endpoints use JWT bearer
    tokens; admin endpoints use session cookies.
  version: v1
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/customer/me/trips/{id}/end:
    post:
      tags:
      - Customer-Trips
      summary: End trip
      description: End the active trip (awaiting dropoff inspection).
      operationId: end
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TripResponse'
...
  /api/admin/trips/{id}/close:
    post:
      tags:
      - Admin-Trips
      summary: Close trip
      description: Close after inspection; computes final charges.
      operationId: close
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TripResponse'
  /api/admin/pricing/rates:
    post:
      tags:
      - Admin-Settings
      summary: Create rate
      description: Add a pricing rate for a vehicle class.
      operationId: createRate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRateRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RateResponse'
  /api/admin/pricing/promos:
    post:
      tags:
      - Admin-Settings
      summary: Create promo
      description: Create a promo code.
      operationId: createPromo
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePromoRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PromoResponse'
...
  /api/customer/me/trips:
    get:
      tags:
      - Customer-Trips
      summary: Trip history
      description: Paginated trip history.
      operationId: history
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageTripResponse'
  /api/customer/me/trips/{id}:
    get:
      tags:
      - Customer-Trips
      summary: Trip detail
      description: A single trip owned by the customer.
      operationId: get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TripResponse'
...
  /api/admin/trips:
    get:
      tags:
      - Admin-Trips
      summary: List trips
      description: Filterable trips list.
      operationId: list_5
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - IN_PROGRESS
          - ENDED_PENDING_INSPECTION
          - CLOSED
      - name: customerId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageTripResponse'
  /api/admin/trips/{id}:
    get:
      tags:
      - Admin-Trips
      summary: Trip detail
      description: A single trip.
      operationId: get_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TripResponse'
...
components:
  schemas:
...
    TripResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bookingId:
          type: string
          format: uuid
        vehicleId:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - IN_PROGRESS
          - ENDED_PENDING_INSPECTION
          - CLOSED
        actualPickupAt:
          type: string
          format: date-time
        actualDropoffAt:
          type: string
          format: date-time
        odometerStart:
          type: integer
          format: int32
        odometerEnd:
          type: integer
          format: int32
        milesDriven:
          type: integer
          format: int32
        finalChargesCents:
          type: integer
          format: int64
...
    CreateRateRequest:
      required:
      - insuranceTier
      - vehicleClassId
      type: object
      properties:
        vehicleClassId:
          type: string
          format: uuid
        dailyRateCents:
          minimum: 0
          type: integer
          format: int64
        deliveryFeeCents:
          minimum: 0
          type: integer
          format: int64
        outOfZoneDropoffFeeCents:
          minimum: 0
          type: integer
          format: int64
        insuranceTier:
          type: string
        insuranceDailyCents:
          minimum: 0
          type: integer
          format: int64
        effectiveFrom:
          type: string
          format: date-time
    RateResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        vehicleClassId:
          type: string
          format: uuid
        dailyRateCents:
          type: integer
          format: int64
    CreatePromoRequest:
      required:
      - code
      - discountType
      type: object
      properties:
        code:
          type: string
        discountType:
          type: string
        discountValue:
          minimum: 1
          type: integer
          format: int32
        maxUses:
          type: integer
          format: int32
        startsAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
    PromoResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        discountType:
          type: string
        discountValue:
          type: integer
          format: int32
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    PageTripResponse:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/TripResponse'
        number:
          type: integer
          format: int32
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
        offset:
          type: integer
          format: int64
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
    SortObject:
      type: object
      properties:
        direction:
          type: string
        nullHandling:
          type: string
        ascending:
          type: boolean
        property:
          type: string
        ignoreCase:
          type: boolean
...
  securitySchemes:
    bearer-jwt:
      type: http
      description: Customer JWT access token.
      scheme: bearer
      bearerFormat: JWT
    session-cookie:
      type: apiKey
      description: Admin session cookie.
      name: SESSION
      in: cookie